Req 7

+++ src/main/java/com/betrybe/agrix/model/repositories/CropRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * The interface Crop repository.
  */
 @Repository
-public interface CropRepository extends JpaRepository<Crop, Integer> {
+public interface CropRepository extends JpaRepository<Crop, Long> {
 
 }
